Working out conflicting gradle builds, before pr
