+++ app/src/main/java/com/example/myapplication/MainActivity.kt
@@ -26,6 +26,7 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         setContent {
             MyApplicationTheme(dynamicColor = false) {
                 val navController = rememberNavController()
